+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 #### 3.0.3
-###### Not yet released
+###### Thursday March 3, 2016
 
 - Add backward compatibilty to print queue to support print later jobs saved with SDK version prior to 3.x
 - Add support for multiple images in single print job